requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
arq==0.26.3
attrs==25.4.0
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.118.3
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
firecrawl-py==4.3.7
flower==2.0.1
frozenlist==1.8.0
gevent==25.9.1
greenlet==3.2.4
h11==0.16.0
hiredis==3.2.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
humanize==4.13.0
idna==3.10
jinja2==3.1.6
kombu==5.5.4
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
multidict==6.7.0
nest-asyncio==1.6.0
packaging==25.0
pandas==2.2.3
prometheus-client==0.23.1
prompt-toolkit==3.0.52
propcache==0.4.1
pycparser==2.23
pydantic==2.12.0
pydantic-core==2.41.1
pydantic-settings==2.11.0
pygments==2.19.2
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.3
redis==5.3.1
requests==2.32.5
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.0
sentry-sdk==2.41.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
tenacity==9.1.2
tornado==6.5.2
typer==0.19.2
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.14
websockets==15.0.1
yarl==1.22.0
zope-event==6.0
zope-interface==8.0.1
